+++ sdk/storage/Azure.Storage.Files.Shares/src/ShareExtensions.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using System;
+using System.Collections.Generic;
 using System.Globalization;
 using System.Text;
 using Azure.Storage.Files.Shares.Models;
@@ -39,6 +40,21 @@ namespace Azure.Storage.Files.Shares
         private static string ToFileDateTimeString(this DateTimeOffset dateTimeOffset)
             => dateTimeOffset.UtcDateTime.ToString(Constants.File.FileTimeFormat, CultureInfo.InvariantCulture);
 
+        internal static string ToShareEnableProtocolsString(this ShareProtocols? shareEnabledProtocols)
+        {
+            if (shareEnabledProtocols == null)
+            {
+                return null;
+            }
+
+            return shareEnabledProtocols switch
+                {
+                    ShareProtocols.Smb => Constants.File.SmbProtocol,
+                    ShareProtocols.Nfs => Constants.File.NfsProtocol,
+                    _ => throw new ArgumentException($"Unknown share protocol: {shareEnabledProtocols}"),
+                };
+        }
+
         internal static ShareSnapshotsDeleteOptionInternal? ToShareSnapshotsDeleteOptionInternal(this ShareSnapshotsDeleteOption? option)
         {
             if (option == null)
@@ -51,6 +67,100 @@ namespace Azure.Storage.Files.Shares
                 ShareSnapshotsDeleteOption.IncludeWithLeased => ShareSnapshotsDeleteOptionInternal.IncludeLeased,
                 _ => throw new ArgumentException($"Invalid {nameof(ShareSnapshotsDeleteOption)}: {option}"),
             };
+        }
+
+        internal static IEnumerable<ShareItem> ToShareItems(this IEnumerable<ShareItemInternal> shareItemInternals)
+        {
+            if (shareItemInternals == null)
+            {
+                return null;
+            }
+
+            List<ShareItem> result = new List<ShareItem>();
+            foreach (ShareItemInternal shareItemInternal in shareItemInternals)
+            {
+                result.Add(shareItemInternal.ToShareItem());
+            }
+            return result;
+        }
+
+        internal static ShareItem ToShareItem(this ShareItemInternal shareItemInternal)
+        {
+            if (shareItemInternal == null)
+            {
+                return null;
+            }
+
+            return new ShareItem
+            {
+                Name = shareItemInternal.Name,
+                Snapshot = shareItemInternal.Snapshot,
+                IsDeleted = shareItemInternal.IsDeleted,
+                VersionId = shareItemInternal.VersionId,
+                Properties = shareItemInternal.Properties.ToShareProperties()
+            };
+        }
+
+        internal static ShareProperties ToShareProperties(this SharePropertiesInternal sharePropertiesInternal)
+        {
+            if (sharePropertiesInternal == null)
+            {
+                return null;
+            }
+
+            return new ShareProperties
+            {
+                LastModified = sharePropertiesInternal.LastModified,
+                ETag = sharePropertiesInternal.ETag,
+                ProvisionedIops = sharePropertiesInternal.ProvisionedIops,
+                ProvisionedIngressMBps = sharePropertiesInternal.ProvisionedIngressMBps,
+                ProvisionedEgressMBps = sharePropertiesInternal.ProvisionedEgressMBps,
+                NextAllowedQuotaDowngradeTime = sharePropertiesInternal.NextAllowedQuotaDowngradeTime,
+                DeletedOn = sharePropertiesInternal.DeletedOn,
+                RemainingRetentionDays = sharePropertiesInternal.RemainingRetentionDays,
+                AccessTier = sharePropertiesInternal.AccessTier,
+                AccessTierChangeTime = sharePropertiesInternal.AccessTierChangeTime,
+                AccessTierTransitionState = sharePropertiesInternal.AccessTierTransitionState,
+                LeaseStatus = sharePropertiesInternal.LeaseStatus,
+                LeaseState = sharePropertiesInternal.LeaseState,
+                LeaseDuration = sharePropertiesInternal.LeaseDuration,
+                Protocols = ToShareEnabledProtocols(sharePropertiesInternal.EnabledProtocols),
+                RootSquash = sharePropertiesInternal.RootSquash,
+                QuotaInGB = sharePropertiesInternal.QuotaInGB,
+                Metadata = sharePropertiesInternal.Metadata
+            };
+        }
+
+        internal static ShareProtocols? ToShareEnabledProtocols(string rawProtocols)
+        {
+            if (rawProtocols == null)
+            {
+                return null;
+            }
+
+            string[] split = rawProtocols.Split(',');
+
+            int result = 0;
+
+            foreach (string s in split)
+            {
+                switch (s)
+                {
+                    case Constants.File.SmbProtocol:
+                        result |= (int)ShareProtocols.Smb;
+                        break;
+                    case Constants.File.NfsProtocol:
+                        result |= (int)ShareProtocols.Nfs;
+                        break;
+                }
+            }
+
+            if (result == 0)
+            {
+                return null;
+            }
+
+            return (ShareProtocols)result;
         }
     }
 }