+++ sdk/storage/Azure.Storage.Files.Shares/tests/FileTestBase.cs
@@ -19,7 +19,8 @@ namespace Azure.Storage.Files.Shares.Tests
         ShareClientOptions.ServiceVersion.V2019_02_02,
         ShareClientOptions.ServiceVersion.V2019_07_07,
         ShareClientOptions.ServiceVersion.V2019_12_12,
-        ShareClientOptions.ServiceVersion.V2020_02_10)]
+        ShareClientOptions.ServiceVersion.V2020_02_10,
+        ShareClientOptions.ServiceVersion.V2020_04_08)]
     public class FileTestBase : StorageTestBase
     {
         protected readonly ShareClientOptions.ServiceVersion _serviceVersion;