+++ sdk/storage/Azure.Storage.Queues/src/QueueClientOptions.cs
@@ -53,7 +53,12 @@ namespace Azure.Storage.Queues
             /// <summary>
             /// The 2020-02-10 service version.
             /// </summary>
-            V2020_02_10 = 4
+            V2020_02_10 = 4,
+
+            /// <summary>
+            /// The 2020-04-08 service version.
+            /// </summary>
+            V2020_04_08 = 5
 #pragma warning restore CA1707 // Identifiers should not contain underscores
         }