+++ sdk/storage/Azure.Storage.Files.Shares/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release History
 
 ## 12.6.0-beta.1 (Unreleased)
+- Added support for service version 2020-04-08.
+- Added support for Share Enabled Protocol and Share Squash Root.
 - Fixed bug where ShareServiceClient.GetShareClient(), ShareClient.GetDirectoryClient(), ShareClient.GetRootDirectoryClient(), ShareClient.WithSnapshot(), ShareDirectoryClient.GetSubDirectoryClient() and ShareDirectoryClient.GetFileClient() created clients that could not generate a SAS from clients that could generate a SAS
 
 ## 12.5.0 (2020-11-10)